+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - Fixed storage key prefix for SecurityManager 
 - Disabled encrypted shared preferences for Android secure storage, it now uses KeyStore by default
 
+### Developer related
+- Added README's to all components
+
 ## 1.2.4
 
 ### New features